requirements.txt
pandas==2.2.2
openpyxl==3.1.5
PyYAML==6.0.2
python-dotenv==1.0.1
playwright==1.48.0
Typer==0.12.5
loguru==0.7.2
pydantic==2.9.2
xlrd==2.0.1
